feat(require-2fa): restrict removing last 2fa method (#9063)

+++ src/sentry/api/endpoints/user_authenticator_details.py
@@ -133,6 +133,20 @@ class UserAuthenticatorDetailsEndpoint(UserEndpoint):
             )
             return Response(status=status.HTTP_204_NO_CONTENT)
 
+        # if the user's organization requires 2fa,
+        # don't delete the last auth method
+        enrolled_methods = Authenticator.objects.all_interfaces_for_user(user, ignore_backup=True)
+        last_2fa_method = len(enrolled_methods) == 1
+        require_2fa = user.get_orgs_require_2fa().exists()
+
+        if require_2fa and last_2fa_method:
+            return Response(
+                {
+                    'detail': 'Cannot delete authenticator because organization requires 2FA',
+                },
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         with transaction.atomic():
             authenticator.delete()
